requirements.txt
cassandra-driver==3.29.2
click==8.1.8
confluent-kafka==2.10.0
geomet==0.2.1.post1
py4j==0.10.9.7
pyspark==3.5.5
python-dotenv==1.1.0
PyYAML==6.0.2
six==1.17.0
websocket-client==1.8.0
